Initial replacement of RNG
